Split `.rest_framework` module into package. [ci skip]
