+++ django_pydantic_field/v2/rest_framework/coreapi.py
@@ -1,0 +1,5 @@
+from rest_framework.schemas import coreapi
+
+
+class AutoSchema(coreapi.AutoSchema):
+    """Not implemented yet."""
