+++ django_pydantic_field/v2/rest_framework/openapi.py
@@ -1,0 +1,5 @@
+from rest_framework.schemas import openapi
+
+
+class AutoSchema(openapi.AutoSchema):
+    """Not implemented yet."""
